Updated develop and merge into current. Moved methods from auth controller to user controller (get, get all, password).
